+++ src/components/Card/Card.tsx
@@ -8,21 +8,16 @@ import { CardStyled } from "./Card.styled";
 // prettier-ignore
 export interface CardCustomProps {
     shape  ? : "rounded" | "curved";
-    shadow ? : "mild" | "soft" | "hard";
 }
 
 export type CardElementType = HTMLDivElement;
 export type CardProps = Omit<CommonAndHTMLProps<CardElementType>, keyof CardCustomProps> & CardCustomProps;
 
-export const Card = React.forwardRef(({ shadow, shape, ...props }: CardProps, ref: React.Ref<CardElementType>) => {
+export const Card = React.forwardRef(({ shape, ...props }: CardProps, ref: React.Ref<CardElementType>) => {
     let classNames = [];
 
     if (shape) {
         classNames.push(`shape-${shape}`);
-    }
-
-    if (shadow) {
-        classNames.push(`shadow-${shadow}`);
     }
 
     return <Element<CardElementType> as={CardStyled} ref={ref} classNames={classNames} {...props} />;